Remove dependancy on healpy
