+++ utils_intensity_map.py
@@ -1,5 +1,4 @@
 import numpy as np
-import healpy as hp
 import os
 
 
@@ -25,6 +24,14 @@ def angle2moll(theta, phi):
     y = rad * np.sqrt(2.0) * np.sin(angle2)
     
     return x, y
+
+
+
+def extract_rms(intensity_map_normalized):
+
+    rms = np.sqrt(np.mean(intensity_map_normalized**2))
+
+    return rms
 
 
 
@@ -92,24 +99,6 @@ def extract_run_parameters(iex, dataset_params, facility_spec, sys_params, deck_
     return print_line, mean_power_fraction
 
 
-def alms2power_spectrum(alms, LMAX):
-
-    the_modes = np.zeros(LMAX)
-    the_modes_full = np.zeros((LMAX,LMAX+1))
-    for l in range(LMAX):
-        for m in range(l+1):
-            the_modes_full[l,m] = np.real(alms[hp.sphtfunc.Alm.getidx(LMAX, l, m)]*
-                np.conjugate(alms[hp.sphtfunc.Alm.getidx(LMAX, l, m)]))
-            if (m>0):
-                the_modes[l] = the_modes[l] + 2.*the_modes_full[l,m]
-            else:          
-                the_modes[l] = the_modes[l] + the_modes_full[l,m]
-
-    the_modes = the_modes / (4.*np.pi)
-
-    return the_modes
-
-
 
 def alms2rms(real_modes, imag_modes, lmax):
 
@@ -123,19 +112,6 @@ def alms2rms(real_modes, imag_modes, lmax):
 
 
 
-def power_spectrum(intensity_map, LMAX, verbose=True):
-    intensity_map_normalized, avg_power = imap_norm(intensity_map)
-    alms = hp.sphtfunc.map2alm(intensity_map_normalized, lmax=LMAX)
-    power_spectrum = alms2power_spectrum(alms, LMAX)
-
-    if verbose:
-        print("The LLE quoted rms cumulative over all modes is: ", np.sqrt(np.sum(power_spectrum))*100.0, "%")
-    sqrt_power_spectrum = np.sqrt(power_spectrum)
-
-    return sqrt_power_spectrum
-
-
-
 def create_ytrain(pointing_per_cone, pointing_nside, defocus_per_cone, num_defocus, power_per_cone, num_powers):
 
     Y_train = np.hstack((np.array(pointing_per_cone)/(pointing_nside-1), np.array(defocus_per_cone)/(num_defocus-1)))
@@ -146,54 +122,9 @@ def create_ytrain(pointing_per_cone, pointing_nside, defocus_per_cone, num_defoc
 
 
 
-def extract_modes_and_flux(intensity_map, LMAX):
-
-    intensity_map_normalized, avg_flux = imap_norm(intensity_map)
-    real_modes, imag_modes = imap2modes(intensity_map_normalized, LMAX, avg_flux)
-
-    return real_modes, imag_modes, avg_flux
-
-
-
-def imap2modes(intensity_map_normalized, LMAX, avg_power):
-
-    modes_complex = hp.sphtfunc.map2alm(intensity_map_normalized, lmax=LMAX)
-
-    return modes_complex.real, modes_complex.imag
-
-
-
-def modes2imap(real_modes, imag_modes, imap_nside):
-
-    np_complex = np.vectorize(complex)
-    modes_complex = np_complex(real_modes, imag_modes)
-    intensity_map_normalized = hp.alm2map(modes_complex, imap_nside)
-
-    return intensity_map_normalized
-
-
-
 def imap_norm(intensity_map):
 
     avg_flux = np.mean(intensity_map) # average power per steradian (i.e. a flux)
     intensity_map_normalized = intensity_map / avg_flux - 1.0
 
     return intensity_map_normalized, avg_flux
-
-
-
-def change_number_modes(Y_train, avg_powers_all, LMAX):
-
-    num_examples = np.shape(Y_train)[1]
-    Y_train2 = np.zeros((LMAX, num_examples))
-    num_coeff = int(((LMAX + 2) * (LMAX + 1))/2.0)
-    np_complex = np.vectorize(complex)
-    for ie in range(num_examples):
-        # weighting to allow NN to adjust for mean flux
-        #Y_train_real = np.squeeze(Y_train[:,ie] / avg_powers_all[ie])
-        Y_train_complex = np_complex(Y_train_real[:num_coeff], Y_train_real[num_coeff:])
-
-        power_spectrum = alms2power_spectrum(Y_train_complex, LMAX)
-        Y_train2[:,ie] = np.sqrt(power_spectrum)
-
-    return Y_train2
